data_collection/data_frontend/requirements.txt
streamlit==1.52.0

torch==2.10.0.dev20251207+cu126
torchvision==0.25.0.dev20251206+cu126
--index-url https://download.pytorch.org/whl/nightly/cu126

pillow==12.0.0
requests==2.32.5
pandas==2.3.3
numpy==2.2.6
